requirements.txt
yahooquery==2.4.1
pandas==2.2.0
numpy==1.26.4
ta==0.11.0
anthropic==0.18.1
python-dotenv==1.0.1
colorama==0.4.6
